Refactor all resaca internals to common KMP (except ViewModelNewInstanceFactory)
