+++ resacacore/build.gradle.kts
@@ -30,6 +30,7 @@ kotlin {
     }
     sourceSets {
         commonMain.dependencies {
+            implementation(libs.uuid)
             implementation(compose.runtime)
             api(libs.androidx.lifecycle.viewmodel.compose)
             api(libs.androidx.core.bundle)
