+++ resacacore/src/commonMain/kotlin/com/sebaslogen/resaca/core/ComposeLifecycleObservers.kt
@@ -1,0 +1,6 @@
+package com.sebaslogen.resaca.core
+
+import androidx.compose.runtime.Composable
+
+@Composable
+public expect fun ObserveComposableContainerLifecycle(scopedViewModelContainer: ScopedViewModelContainer)
