+++ resacacore/src/commonMain/kotlin/com/sebaslogen/resaca/core/PlatformLifecycleHandler.kt
@@ -2,7 +2,7 @@ package com.sebaslogen.resaca.core
 
 public const val COMPOSITION_RESUMED_TIMEOUT_IN_SECONDS: Long = 1
 
-expect public class PlatformLifecycleHandler {
+expect public class PlatformLifecycleHandler() {
     public fun onResumed()
     public fun onDestroyed()
     public suspend fun awaitBeforeDisposing(inForeground: Boolean)
