+++ resacacore/src/commonMain/kotlin/com/sebaslogen/resaca/core/ViewModelNewInstanceFactory.kt
@@ -1,4 +1,4 @@
-package com.sebaslogen.resaca
+package com.sebaslogen.resaca.core
 
 import androidx.lifecycle.ViewModelProvider
 
@@ -6,5 +6,5 @@ import androidx.lifecycle.ViewModelProvider
  * Singleton to mimic [ViewModelProvider.NewInstanceFactory] instance in that library.
  */
 internal object ViewModelNewInstanceFactory {
-    val instance: ViewModelProvider.Factory = ViewModelProvider.NewInstanceFactory()
+    val instance: ViewModelProvider.Factory = TODO() //ViewModelProvider.NewInstanceFactory()
 }
