+++ resacacore/src/iosMain/kotlin/com/sebaslogen/resaca/core/ComposeLifecycleObservers.ios.kt
@@ -1,0 +1,7 @@
+package com.sebaslogen.resaca.core
+
+import androidx.compose.runtime.Composable
+
+@Composable
+actual fun ObserveComposableContainerLifecycle(scopedViewModelContainer: ScopedViewModelContainer) {
+}
