+++ resacahilt/src/main/java/com/sebaslogen/resaca/hilt/ScopedMemoizers.kt
@@ -15,13 +15,13 @@ import androidx.lifecycle.ViewModelStoreOwner
 import androidx.lifecycle.viewmodel.CreationExtras
 import androidx.lifecycle.viewmodel.compose.LocalViewModelStoreOwner
 import androidx.navigation.NavBackStackEntry
-import com.sebaslogen.resaca.ScopedViewModelContainer
-import com.sebaslogen.resaca.ScopedViewModelContainer.ExternalKey
-import com.sebaslogen.resaca.ScopedViewModelContainer.InternalKey
-import com.sebaslogen.resaca.ScopedViewModelOwner
+import com.sebaslogen.resaca.core.ScopedViewModelContainer
+import com.sebaslogen.resaca.core.ScopedViewModelContainer.ExternalKey
+import com.sebaslogen.resaca.core.ScopedViewModelContainer.InternalKey
+import com.sebaslogen.resaca.core.ScopedViewModelOwner
 import com.sebaslogen.resaca.core.KeyInScopeResolver
 import com.sebaslogen.resaca.core.ScopeKeyWithResolver
-import com.sebaslogen.resaca.generateKeysAndObserveLifecycle
+import com.sebaslogen.resaca.core.generateKeysAndObserveLifecycle
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.hilt.android.AndroidEntryPoint
