+++ resacakoin/src/main/java/com/sebaslogen/resaca/koin/ScopedMemoizers.kt
@@ -6,13 +6,13 @@ import androidx.compose.runtime.remember
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.ViewModelStore
-import com.sebaslogen.resaca.ScopedViewModelContainer
-import com.sebaslogen.resaca.ScopedViewModelContainer.ExternalKey
-import com.sebaslogen.resaca.ScopedViewModelContainer.InternalKey
-import com.sebaslogen.resaca.ScopedViewModelOwner
+import com.sebaslogen.resaca.core.ScopedViewModelContainer
+import com.sebaslogen.resaca.core.ScopedViewModelContainer.ExternalKey
+import com.sebaslogen.resaca.core.ScopedViewModelContainer.InternalKey
+import com.sebaslogen.resaca.core.ScopedViewModelOwner
 import com.sebaslogen.resaca.core.KeyInScopeResolver
 import com.sebaslogen.resaca.core.ScopeKeyWithResolver
-import com.sebaslogen.resaca.generateKeysAndObserveLifecycle
+import com.sebaslogen.resaca.core.generateKeysAndObserveLifecycle
 import org.koin.androidx.viewmodel.factory.KoinViewModelFactory
 import org.koin.core.annotation.KoinInternalApi
 import org.koin.core.context.GlobalContext
