+++ sample/src/main/java/com/sebaslogen/resacaapp/sample/ui/main/compose/examples/SimpleViewModel.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.SavedStateViewModelFactory
 import androidx.lifecycle.ViewModel
-import com.sebaslogen.resaca.viewModelScoped
+import com.sebaslogen.resaca.core.viewModelScoped
 import com.sebaslogen.resacaapp.sample.ui.main.compose.DemoComposable
 import com.sebaslogen.resacaapp.sample.ui.main.data.FakeScopedViewModel
