+++ sample/src/test/java/com/sebaslogen/resacaapp/sample/ScopeKeysTest.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.test.hasTextExactly
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.performClick
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.sebaslogen.resaca.rememberScoped
+import com.sebaslogen.resaca.core.rememberScoped
 import com.sebaslogen.resacaapp.sample.ui.main.compose.DemoComposable
 import com.sebaslogen.resacaapp.sample.ui.main.data.FakeRepo
 import com.sebaslogen.resacaapp.sample.utils.ComposeTestUtils
